requirements.txt
djangorestframework~=3.12.2Django~=3.1.7django-enumfields~=2.1.1channels~=3.0.3redis~=3.5.3aioredis~=1.3.1asyncio~=3.4.3